third_party/WebKit/LayoutTests/http/tests/devtools/local-object-properties-section.html

<html>
<head>
<script src="../inspector/inspector-test.js"></script>
<script>

function initialize_TestObject() {
    InspectorTest.preloadModule("object_ui");
}

function test() {
  var d = [];
  for (var i = 1000; i < 1256; ++i)
    d.push(i);
  var object = {a: 'b', c: d};
  var localObject = SDK.RemoteObject.fromLocalObject(object);

  var propertiesSection = new ObjectUI.ObjectPropertiesSection(localObject, 'local object');
  propertiesSection.expand();
  propertiesSection.objectTreeElement().childAt(1).expand();

  TestRunner.addResult(TestRunner.textContentWithLineBreaks(propertiesSection.element));
  TestRunner.completeTest();
}

</script>
</head>

<body onload="runTest()">
<p>Test that ObjectPropertiesSection works with local remote objects.</p>
</body>
</html>
