third_party/WebKit/LayoutTests/http/tests/devtools/local-object-properties-section-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Test that ObjectPropertiesSection works with local remote objects.

local object
    a: "b"
    c: [1000, 1001, 1002, 1003, 1004, 1005, 1006, 1007, 1008, 1009, 1010, 1011, 1012, 1013, 1014, 1015, 1016,…]
        [0 … 99]
        [100 … 199]
        [200 … 255]
